refactor(fs): unify FsStatsLike into fs/types.ts
